+++ src/main/java/url/UrlRequest.java
@@ -1,4 +1,4 @@
-package urlshortener;
+package url;
 
 public record UrlRequest(String orignalUrl) {
 }
