+++ src/main/java/url/UrlResponse.java
@@ -1,4 +1,4 @@
-package urlshortener;
+package url;
 
 public record UrlResponse(String shortUrl) {
 }
